move actions and reduser for film to own directory
